+++ gs_capture_addon/core/training/nerfstudio.py
@@ -9,7 +9,7 @@ import shutil
 from typing import Optional, List, Dict
 
 from .base import TrainingBackend, TrainingConfig, TrainingProgress, TrainingStatus
-from ...utils.paths import normalize_path, get_conda_base, get_conda_script, get_conda_executable
+from ...utils.paths import normalize_path, get_conda_script, get_conda_executable
 
 
 class NerfstudioBackend(TrainingBackend):